+++ server/src/main/java/pl/edu/agh/server/student/StudentService.java
@@ -8,6 +8,7 @@ import java.util.List;
 public class StudentService {
 
     private final StudentRepository repository;
+
     public StudentService(StudentRepository repository) {
         this.repository = repository;
     }
@@ -15,24 +16,36 @@ public class StudentService {
     public void addStudent(Student student) {
         repository.save(student);
     }
+
     public List<Student> getStudents() {
         return repository.findAll();
     }
+
     public Student getStudent(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Student with id: " + id + "not found"));
+        try {
+            return repository.findById(id).orElseThrow(() -> new RuntimeException("Student with id: " + id + "not found"));
+        } catch (RuntimeException e) {
+            System.out.println(e.getMessage());
+            return null;
+        }
     }
+
     public Student save(Student student) {
         return repository.save(student);
     }
+
     public List<Student> saveMany(List<Student> students) {
         return repository.saveAll(students);
     }
+
     public void deleteStudent(Long id) {
         repository.deleteById(id);
     }
+
     public void deleteStudent(Student student) {
         repository.delete(student);
     }
+
     public void updateStudent(Student student) {
         repository.save(student);
     }
