+++ server/src/main/java/pl/edu/agh/server/term/TermService.java
@@ -1,7 +1,6 @@
 package pl.edu.agh.server.term;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -11,7 +10,7 @@ import java.util.List;
 public class TermService {
     private final TermRepository termRepository;
 
-    public List<Term> getTerms(){
+    public List<Term> getTerms() {
         return termRepository.findAll();
     }
 
@@ -20,7 +19,7 @@ public class TermService {
                 "Term  with ID " + termId + " does not exist"));
     }
 
-    public void createNewTerms(List<Term> terms){
+    public void createNewTerms(List<Term> terms) {
         termRepository.saveAll(terms);
     }
 }
